♻️ Convert to using modern gem defaults
This is roughly the result of running `bundle gem staccato`, merging
and configuring that result with the existing project, and adding a few
binstubs.  It adds rubocop, GitHub Actions, and rbs.

The required ruby version is raised from 2.0 to 2.6, so that a (mostly)
up-to-date rubocop can be run—all 2.x versions of ruby are EOL and this
simplifies requirements for running rubocop and other CI tools.  This
also simplifies any future PRs, by allowing them to take advantage of
modern ruby features.

In addition to the default rubocop config provided by the `bundle gem`
template, the base baseline rubocop config inherits from standard gem.
Contrary to the "standard" propaganda, it's okay to inherit from,
disable, and override the "standard" config.  😉  Its config is still
far better than the rubocop defaults, and a better fit for the
pre-existing style used in this project.  Where this project
significantly deviates from that "standard" config, I used my personal
judgement about which rules to disable or configure in `rubocop.yml` and
which ones to generate into `rubocop_todo.yml`.

The Travis config was replaced with GitHub Actions, running on every
supported version of ruby.

The rbs file was simply generated by typeprof, and I didn't audit the
generated rbs for correctness or completeness.